Shoot functonailty for sphere intersection seems to be working. But the shading is not. This commit makes correctly, and runs, but does not shade.

+++ linkedlist.h
@@ -7,7 +7,7 @@
 typedef struct pixel {
 
 	// The is where the colors for a pixel are stored
-    int R, G, B;
+    float R, G, B;
 
 } pixel;
